+++ src/test/java/com/bookspot/batch/openapi/LibraryApiRequesterTest.java
@@ -1,8 +1,7 @@
 package com.bookspot.batch.openapi;
 
-import com.bookspot.batch.data.library.Library;
+import com.bookspot.batch.library.data.Library;
 import com.bookspot.batch.library.api.LibraryApiRequester;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.PageRequest;
